moved cal-bot into a subfolder
