+++ .gitignore
@@ -3,7 +3,7 @@ build
 npm-debug.log
 
 # don't commit the file that simulates the AWS environment variables
-dev/.env
+cla-bot/dev/.env
 
 # I don't want to commit the private RSA keys for my bot!
 # see: https://docs.travis-ci.com/user/encrypting-files/#Encrypting-multiple-files
