+++ cla-bot/dev/.env
@@ -1,0 +1,6 @@
+GITHUB_ACCESS_TOKEN=8e666890184a1042634684a3f0e765be90591347
+INTEGRATION_ENABLED=false
+INTEGRATION_KEY=clabot-dev-integration-key.pem
+INTEGRATION_ID=2208
+LOG_URL=https://colineberhardt.github.io/cla-bot/
+BOT_NAME=cla-bot
