+++ spec/lambdaSpec.js
@@ -123,7 +123,7 @@ describe('lambda function', () => {
 
   it('should ignore actions that are not pull requests being opened', (done) => {
     event.body.action = 'label';
-    const lambda = require('../index');
+    const lambda = require('../cla-bot/index');
     lambda.handler(event, {}, (err, result) => {
       expect(err).toBeNull();
       expect(result.message).toEqual('ignored action of type label');
@@ -135,7 +135,7 @@ describe('lambda function', () => {
     xit('should propagate HTTP request errors', (done) => {
       // create a mal-formed URL
       event.body.repository.url = 'http:://foo.com/user/repo';
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
       lambda.handler(event, {}, (err) => {
         expect(err).toEqual('Error: Invalid URI "http:://foo.com/user/repo/contents/.clabot"');
         done();
@@ -159,7 +159,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err) => {
@@ -183,7 +183,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err) => {
@@ -207,7 +207,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err) => {
@@ -231,7 +231,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err) => {
@@ -260,7 +260,7 @@ describe('lambda function', () => {
       ], installationToken));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {}, done);
     });
@@ -274,7 +274,7 @@ describe('lambda function', () => {
       ], installationToken));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {}, done);
     });
@@ -304,7 +304,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err, result) => {
@@ -337,7 +337,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err, result) => {
@@ -360,7 +360,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err, result) => {
@@ -394,7 +394,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err) => {
@@ -437,7 +437,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err, result) => {
@@ -473,7 +473,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err, result) => {
@@ -513,7 +513,7 @@ describe('lambda function', () => {
       }));
 
       mock('request', request);
-      const lambda = require('../index');
+      const lambda = require('../cla-bot/index');
 
       lambda.handler(event, {},
         (err, result) => {
@@ -551,7 +551,7 @@ describe('contributionVerifier', () => {
       });
 
       mock('request', request);
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob'])
         .then((nonContributors) => {
@@ -572,7 +572,7 @@ describe('contributionVerifier', () => {
       });
 
       mock('request', request);
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob', 'billy'])
         .then((nonContributors) => {
@@ -600,7 +600,7 @@ describe('contributionVerifier', () => {
       });
 
       mock('request', request);
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob', 'foo'])
         .then((nonContributors) => {
@@ -616,7 +616,7 @@ describe('contributionVerifier', () => {
         contributors: ['billy']
       };
 
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob', 'billy'])
         .then((nonContributors) => {
@@ -637,7 +637,7 @@ describe('contributionVerifier', () => {
       });
 
       mock('request', request);
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob', 'billy'])
         .then((nonContributors) => {
@@ -663,7 +663,7 @@ describe('contributionVerifier', () => {
       });
 
       mock('request', request);
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob'])
         .then((nonContributors) => {
@@ -691,7 +691,7 @@ describe('contributionVerifier', () => {
       });
 
       mock('request', request);
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
 
       verifier(config)(['bob', 'foo'])
         .then((nonContributors) => {
@@ -702,7 +702,7 @@ describe('contributionVerifier', () => {
 
     it('should throw an error if no configuration is supplied', () => {
       const config = {};
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
       expect(() => verifier(config))
         .toThrow(new Error('A mechanism for verifying contributors has not been specified'));
     });
@@ -711,7 +711,7 @@ describe('contributionVerifier', () => {
       const config = {
         contributors: 'sausage'
       };
-      const verifier = require('../contributionVerifier');
+      const verifier = require('../cla-bot/contributionVerifier');
       expect(() => verifier(config))
         .toThrow(new Error('A mechanism for verifying contributors has not been specified'));
     });
@@ -719,7 +719,7 @@ describe('contributionVerifier', () => {
 });
 
 describe('lambda internals', () => {
-  const internals = require('../index').test;
+  const internals = require('../cla-bot/index').test;
 
   describe('commentSummonsBot', () => {
     it('should match comments with one space', () => {
